requirements.txt
boto3==1.24.66
dataiter==0.49
entsoe-py==0.6.2
flake8==5.0.4
numpy==1.23.2
pandas==1.4.4
requests==2.28.1
